FIS-89 Create packages for tests
