+++ src/test/java/ClientTest/TenRequestsExceptionTest.java
@@ -1,3 +1,5 @@
+package ClientTest;
+
 import Client.Exception.TenRequestsException;
 import org.junit.Test;
 import static org.junit.Assert.*;
