+++ src/test/java/ParentTest/DecodeTest.java
@@ -1,3 +1,5 @@
+package ParentTest;
+
 import Manager.Services.AddJSON;
 import ParentCode.Decode;
 import ParentCode.Exceptions.AlreadyExistsException;
