+++ src/test/java/ParentTest/ServiceTest.java
@@ -1,3 +1,5 @@
+package ParentTest;
+
 import Register.services.FileSystemService;
 import Register.services.UserService;
 import org.apache.commons.io.FileUtils;
